Update WinRM config options
A new config group is being added, allowing specifying WinRM
credentials to be used when accessing Hyper-V hosts. This will
also allow setting a transport type other than 'plaintext'.

At the same time, this change adds a new helper method that may
be used in order to run commands on the Hyper-V hosts.

Change-Id: I5852154d8275ebd36785bcf56899da3257f1a5a8