Adds metrics collection scenario
This test suite verifies that the instance metrics are properly published
and collected and have non-zero values. The verification is done via the
ceilometer API.

Waiting for the ceilometer compute agent to poll the resources is crucial,
otherwise the test suite will fail due to the fact that no samples would
be found published before checking the samples.

The ceilometer agent's polling interval should have the same value as
the test suite's polled_metric_delay. This can be done in two ways:
a. Configure tempest's polled_metric_delay, by adding the following
line in tempest.conf, in the telemetry section:
    polled_metric_delay = <desired value>
b. Set the interval value in pipeline.yaml on the compute node to the
desired value and restart the ceilometer compute agent.

The second method is preferred, as the interval value defined in
pipeline.yaml is 600 seconds, which would mean each test would last
at least 10 minutes.

Change-Id: I2ad696c221fba860a30621686b464b8486bea583